session.c: silence compiler warning
gcc
```
src/session.c: In function ‘libssh2_poll’:
src/session.c:1550:33: warning: conversion from ‘long unsigned int’ to ‘short int’ may change value [-Wconversion]
 1550 |             sockets[i].events = fds[i].events;
      |                                 ^~~
```

clang
```
src/session.c:1550:40: warning: implicit conversion loses integer precision: 'unsigned long' to 'short' [-Wimplicit-int-conversion]
            sockets[i].events = fds[i].events;
                              ~ ~~~~~~~^~~~~~
```

+++ src/session.c
@@ -1547,7 +1547,7 @@ libssh2_poll(LIBSSH2_POLLFD * fds, unsigned int nfds, long timeout)
         switch(fds[i].type) {
         case LIBSSH2_POLLFD_SOCKET:
             sockets[i].fd = fds[i].fd.socket;
-            sockets[i].events = fds[i].events;
+            sockets[i].events = (short)fds[i].events;
             sockets[i].revents = 0;
             break;
